chore: changed directory structure of helper functions and constants. added video formats to list
