+++ constants.py
@@ -1,0 +1,21 @@
+MEDIA_FILE_TYPES: list[str] = (
+    ".jpg",
+    ".jpeg",
+    ".png",
+    ".gif",
+    ".bmp",
+    ".tiff",
+    ".svg",
+    ".webp",
+    ".heic",
+    ".heif",
+    ".raw",
+    ".nef",
+    ".mov",
+    ".mp4",
+    ".avi",
+    ".mkv",
+    ".wmv",
+    ".flv",
+    ".m4v",
+)
